software: extract test code to test_*.py files.
