+++ software/glasgow/gateware/lfsr.py
@@ -51,33 +51,3 @@ class LinearFeedbackShiftRegister(Elaboratable):
             value = ((value << 1) & mask) | feedback
             if value == self.reset:
                 break
-
-# -------------------------------------------------------------------------------------------------
-
-import unittest
-
-from . import simulation_test
-
-
-class LFSRTestbench(Elaboratable):
-    def __init__(self, **kwargs):
-        self.dut = LinearFeedbackShiftRegister(**kwargs)
-
-    def elaborate(self, platform):
-        return self.dut
-
-
-class LFSRTestCase(unittest.TestCase):
-    def setUp(self):
-        self.tb = LFSRTestbench(degree=16, taps=(16, 14, 13, 11))
-
-    @simulation_test
-    def test_generate(self, tb):
-        soft_values = list(self.tb.dut.generate())
-        hard_values = []
-        for _ in range(len(soft_values)):
-            hard_values.append((yield self.tb.dut.value))
-            yield
-
-        self.assertEqual(len(soft_values), 65535)
-        self.assertEqual(hard_values, soft_values)
